Add namespace to binding filter, refactor
